[fix] exit

+++ libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: admin <admin@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/06 13:31:15 by vserra            #+#    #+#             */
-/*   Updated: 2021/09/22 01:11:31 by admin            ###   ########.fr       */
+/*   Updated: 2021/11/11 14:46:24 by admin            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long	ft_atoi(const char *str)
+double long	ft_atoi(const char *str)
 {
-	long	i;
-	long	sign;
-	long	nbr;
+	long		i;
+	long		sign;
+	double long	nbr;
 
 	i = 0;
 	sign = 1;
